refactor : 중복된 Member 엔티티 제거 및 통합
