--- src/main/java/goormthon_group4/backend/domain/user/entity/Member.java
@@ -1,29 +1,0 @@
-package goormthon_group4.backend.domain.user.entity;
-
-import goormthon_group4.backend.domain.team.entity.Team;
-import goormthon_group4.backend.global.common.base.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-// User는 여러 Team에 참여 가능, 한 team에도 여러 User 참여 가능
-// Member 테이블이 N:M 분리
-public class Member extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "team_id", nullable = false)
-    private Team team;
-
-}
